% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_profile_with_mass.R
\name{ComputeMatrixProfile}
\alias{ComputeMatrixProfile}
\title{ComputeMatrixProfile}
\usage{
ComputeMatrixProfile(data, n)
}
\arguments{
\item{data}{Refere-se a um vetor numérico representando a série temporal}

\item{n}{se refere ao tamanho da subsequência desejada, a mesma utilizada para 
buscar por discords ou motifs}
}
\value{
Retorna um data.frame contendo os valores calculados para a matrix profile
e o índice correspondente a cada distância
}
\description{
Realiza o cálculo da matrix profile para uma determinada série temporal
}
\examples{
ts <- runif(n = 1000, min = 0, max = 100)
subLen <- 10
ComputeMatrixProfile(ts,subLen)
}
