% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_prediction.R
\name{create.event.predictor}
\alias{create.event.predictor}
\title{Set up a Predictor of Binary Events Based on Repeated Measurements}
\usage{
create.event.predictor(rmp, dataset, covariates, event.name = "event",
  verbose = F)
}
\arguments{
\item{rmp}{The results of a call to make.measurement.predictor}

\item{dataset}{A data frame with rows $time, $event (1 or 0), $id}

\item{event.name}{Specifying a name allows for more than one event to be predicted}
}
