% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmp_master.R
\name{create.measurement.predictor}
\alias{create.measurement.predictor}
\title{Create a Measurement Predictor}
\usage{
create.measurement.predictor(dataset, num.latent.classes = 1,
  time.transformation, test.minima, test.maxima,
  time.transformation.parameters = NULL, fixed.covariate.names = NULL,
  initial.class.assignments = NULL, iterations.after.burn, burn, keep,
  chains = 2, pi.alpha = 1, seed = 1223)
}
\arguments{
\item{dataset}{A data frame with the columns 'measurement', 'test', 'id', 'time', as well as column for each entry in 'fixed.covariate.names'}

\item{time.transformation}{One of 'linear', 'linear.spline'}
}
