% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_profile_with_mass.R
\name{FindDiscords}
\alias{FindDiscords}
\title{FindDiscords}
\usage{
FindDiscords(mp.current, subs.length, num_discords = 3)
}
\arguments{
\item{mp.current}{Refere-se a matrix profile atual, a qual se deseja buscar pelos discords}

\item{subs.length}{se refere ao tamanho do discord que se deseja buscar (mesmo tamanho de
subsequência usado em \code{ComputeMatrixProfile})}

\item{num_discords}{quantidade de discords que se deseja encontrar}
}
\value{
Retorna vetor indicando os índices das subsequências onde os discords foram encontrados
}
\description{
Extrai os discords a partir da matrix profile
}
\examples{
ts <- runif(n = 1000, min = 0, max = 100)
subLen <- 10
mp <- ComputeMatrixProfile(ts,subLen)
motifs <- FindMotifs(ts,subLen,mp)
mp.current <- motifs$matrix.profile
FindDiscords(mp.current,subLen)
}
