% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_profile_with_mass.R
\name{FindMotifs}
\alias{FindMotifs}
\title{FindMotifs}
\usage{
FindMotifs(data, subs.length, matrix.profile, num_motifs = 3,
  num_neighbors = 10, radius = 2)
}
\arguments{
\item{data}{Refere-se a um vetor numérico representando a série temporal}

\item{subs.length}{se refere ao tamanho do motif que se deseja buscar (mesmo tamanho de
subsequência usado em \code{ComputeMatrixProfile})}

\item{matrix.profile}{refere-se a matrix profile previamente calculada pelo método
\code{ComputeMatrixProfile}}

\item{num_motifs}{quantidade de motifs que se deseja encontrar}

\item{num_neighbors}{refere-se a quantidade máxima de vizinhos a se
considerar por motifs}

\item{radius}{refere-se ao raio para considerar se uma subsequência caso com 
determinado motif}
}
\value{
Retorna um lista contendo a matrix profile com as zonas de exclusão e 
a lista contendo os índices dos motifs
}
\description{
Extrai os motifs a partir da matrix profile
}
\examples{
ts <- runif(n = 1000, min = 0, max = 100)
subLen <- 10
mp <- ComputeMatrixProfile(ts,subLen)
motifs <- FindMotifs(ts,subLen,mp)
motifs$motifs
}
