% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get.all.class.probabilities}
\alias{get.all.class.probabilities}
\title{Get probabilites of being in each latent class}
\usage{
get.all.class.probabilities(dm, observations, observed.times, covariates)
}
