% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, world!
This is an example function named 'hello'
which prints 'Hello, name'.}
\usage{
hello(name)
}
\arguments{
\item{name}{string}
}
\description{
Hello, world!
This is an example function named 'hello'
which prints 'Hello, name'.
}
