% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_prediction.R
\name{make.event.predictions}
\alias{make.event.predictions}
\title{Make Predictions About Binary Events Based on Repeated Measurements}
\usage{
make.event.predictions(rmp, observations, observed.times, predict.times,
  covariates, event.name = "event", verbose = F)
}
\arguments{
\item{observations}{an array indexed [i,time,test] - where i denotes an individual}

\item{observed.times}{a numeric vector such that observed.times[t] is the time at which observations[,t,] were taken}
}
