% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_profile_with_mass.R
\name{MASS}
\alias{MASS}
\title{MASS}
\usage{
MASS(query, data.length, subs.length, ind, masspre)
}
\arguments{
\item{query}{se a subsequência que se deseja construir a distance profile}

\item{data.length}{se refere ao tamanho da série temporal}

\item{subs.length}{se refere ao tamanho da subsequência desejada, a mesma utilizada para 
buscar por discords ou motifs}

\item{ind}{refere-se a um inteiro que corresponde a trecho que essa subsequência foi
retirada}

\item{masspre}{refere-se ao resultado obtido atrabés da computação do \code{MASSPre}}
}
\value{
Retorna um vetor numérico contendo a distance profile para aquela query
}
\description{
Realiza o cálculo da distance profile para uma determina subsequência
}
\examples{
tsLen <- 100
ts <- runif(n = 1000, min = 0, max = tsLen)
subLen <- 10
masspre <- MASSPre(ts,tsLen,subLen)
index <- 1
query <- ts[index:(index + subLen - 1)]
MASS(query,tsLen,subLen,index,masspre)
}
