% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_profile_with_mass.R
\name{MASSPre}
\alias{MASSPre}
\title{MASSpre}
\usage{
MASSPre(data.values, data.length, subsequence.length)
}
\arguments{
\item{data.values}{se refere a um vetor numérico representando a série temporal}

\item{data.length}{se refere ao tamanho da série temporal passada}

\item{subsequence.length}{se refere ao tamanho da subsequência desejada, a mesma utilizada para 
buscar por discords ou motifs}
}
\value{
Retorna uma lista contendo os dados no domínio na frequência, a média e o desvio padrão
pré-calculados das subsequências
}
\description{
Realiza a pré-computação do MASS, calculando a média móvel e o desvio padrão móvel (movstd,movmean)
}
\examples{
tsLen <- 100
ts <- runif(n = 1000, min = 0, max = tsLen)
MASSPre(ts,1000,10)
}
